+++ constants/addresses.ts
@@ -1,0 +1,1 @@
+export const USDC_ADDRESS = '';
